refactor(tonk-schema): split effect_query into rule_query + effect_query
effect_query mixed two concerns: the rule: query (AnonymousRuleQuery,
RuleDefinition, rule_of_rule_descriptor) that concept and builtin
dispatch into, and the effect storage / lookup / validation
machinery the induce fixpoint uses.

The rule: query moves to rule_query.rs — it stays in tonk-schema
beside concept. effect_query.rs keeps the effect-evaluation items.
The schema definition layer no longer imports effect_query, so it
can later move to a tonk-evaluator crate without a cycle.
