+++ rust/tonk-schema/src/effect_query.rs
@@ -1,31 +1,30 @@
-//! Query and resolution machinery for effects.
+//! Effect storage, lookup, and install-time validation.
 //!
 //! [`effect`](crate::effect) holds the pure [`Effect`] data type
 //! and its storage-shape projections. This module holds the
 //! interpretation side: loading effects back from a branch,
-//! validating the V1 transient-trigger requirement, and the
-//! [`AnonymousRuleQuery`] that surfaces installed rules as
-//! [`ConceptConclusion`] rows.
+//! validating the V1 transient-trigger requirement, and writing
+//! effects into a branch transaction.
+//!
+//! This is the effect-lookup machinery the `induce` fixpoint
+//! uses. It is an upper layer over the schema definition modules
+//! (`concept`, `query_source`, …) — it may freely depend on them.
+//! The schema definition layer must NOT depend on this module;
+//! the `rule:` query that surfaces installed rules as concept
+//! rows lives separately in [`rule_query`](crate::rule_query).
 //!
 //! Splitting this out keeps `effect.rs` a leaf module that only
 //! depends on `dialog-*` crates, so the operation types it
 //! defines can move into their own crate later.
 
-use dialog_artifacts::{Attribute as ArtifactsAttribute, Entity, Select, Value};
+use dialog_artifacts::{Attribute as ArtifactsAttribute, Entity, Value};
 use dialog_capability::{Fork, Provider};
 use dialog_common::ConditionalSync;
 use dialog_effects::archive::{Get, Put};
 use dialog_effects::authority::Identify;
 use dialog_effects::memory::Resolve;
-use dialog_query::concept::descriptor::ConceptConclusion;
-use dialog_query::concept::query::ConceptQuery;
-use dialog_query::source::SelectRules;
-use dialog_query::{
-    Application, ConceptDescriptor, EvaluationError, Match, Output as _, Parameters, Selection,
-    Term, the, try_stream,
-};
+use dialog_query::{Output as _, Term};
 use dialog_repository::{Branch, RemoteSite};
-use serde::{Deserialize, Serialize};
 use thiserror::Error;
 
 use crate::effect::{Effect, EffectError, EffectPolarity};
@@ -265,232 +264,8 @@ impl EffectsByOn {
 }
 
 // ---------------------------------------------------------------- //
-// AnonymousRuleQuery — yields one row per installed rule.          //
+// Writing effects into a branch transaction.                       //
 // ---------------------------------------------------------------- //
-
-/// JSON-serialisable view of a rule's definition — the value of an
-/// [`AnonymousRuleQuery`] row's synthesised `definition` field.
-///
-/// Pairs the rule's [`InductiveRuleDescriptor`](dialog_query::InductiveRuleDescriptor)
-/// (head conclusion, `when` premises, `unless` premises) with its
-/// [`EffectPolarity`] so a `rule:` query surfaces what the rule
-/// does, mirroring how `concept:` puts a concept's descriptor in
-/// its `source` field.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RuleDefinition {
-    /// The inductive rule descriptor: head concept + premises.
-    pub rule: dialog_query::InductiveRuleDescriptor,
-    /// Whether the head asserts or retracts when the body matches.
-    pub polarity: EffectPolarity,
-}
-
-impl RuleDefinition {
-    /// Build the definition view from an installed [`Effect`].
-    fn from_effect(effect: &Effect) -> Self {
-        Self {
-            rule: effect.descriptor(),
-            polarity: effect.polarity(),
-        }
-    }
-}
-
-/// Custom query application that surfaces every inductive rule
-/// installed on a branch as a [`ConceptConclusion`].
-///
-/// Rules are reified as `dialog.effect/*` facts (see the
-/// [`effect`](crate::effect) module docs). This query enumerates
-/// every entity carrying the `dialog.meta/effect = db:effect`
-/// marker, rehydrates each via [`EffectByEntity`], and
-/// materialises its definition into a synthesised `definition`
-/// field alongside `this` (the effect entity). It is the
-/// rule-side parallel of
-/// [`AnonymousConceptQuery`](crate::concept::AnonymousConceptQuery):
-/// same [`Application`] shape, same `terms` filter/emit
-/// convention.
-///
-/// `terms` parameter keys map to the user's variable names:
-/// `this` (the effect entity) and `definition` (the
-/// [`RuleDefinition`] as a JSON string). Constant terms filter;
-/// variable terms bind.
-#[derive(Debug, Clone)]
-pub struct AnonymousRuleQuery {
-    /// Term bindings — keys `this` (effect entity) and
-    /// `definition` (rule definition as a JSON string).
-    pub terms: Parameters,
-}
-
-impl AnonymousRuleQuery {
-    /// Construct a new query from a parameter map.
-    pub fn new(terms: Parameters) -> Self {
-        Self { terms }
-    }
-}
-
-/// Pull a constant entity out of a term — either from a constant
-/// term or from a variable the upstream selection already bound.
-/// Returns `None` if the term is unbound or absent. Mirrors the
-/// `resolve_entity_filter` helper in [`crate::concept`].
-fn resolve_entity_filter(term: &Option<Term<dialog_query::Any>>, input: &Match) -> Option<Entity> {
-    let t = term.as_ref()?;
-    match t {
-        Term::Constant(value) => Entity::try_from(value.clone()).ok(),
-        Term::Variable { name: Some(_), .. } => {
-            let value = input.lookup(t).ok()?;
-            Entity::try_from(value).ok()
-        }
-        Term::Variable { name: None, .. } => None,
-    }
-}
-
-impl Application for AnonymousRuleQuery {
-    type Conclusion = ConceptConclusion;
-
-    fn evaluate<'a, Env, M: Selection + 'a>(self, selection: M, env: &'a Env) -> impl Selection + 'a
-    where
-        Env: Provider<Select<'a>> + Provider<SelectRules> + ConditionalSync,
-    {
-        let app = self;
-        try_stream! {
-            for await each in selection {
-                let input = each?;
-
-                let this_term = app.terms.get("this").cloned();
-                let definition_term = app.terms.get("definition").cloned();
-
-                let this_filter = resolve_entity_filter(&this_term, &input);
-
-                // Enumerate every effect entity via the
-                // `dialog.meta/effect = db:effect` marker. When
-                // `this` is a constant the selection is already
-                // narrowed to that entity.
-                let marker = effect_marker_entity();
-                let this_term_for_marker: Term<Entity> = match &this_filter {
-                    Some(e) => Term::Constant(Value::Entity(e.clone())),
-                    None => Term::var("__rule_query_this"),
-                };
-                let claims: Vec<dialog_query::Claim> = the!("dialog.meta/effect")
-                    .of(this_term_for_marker)
-                    .is(marker)
-                    .perform(env)
-                    .try_vec()
-                    .await?;
-
-                for claim in claims {
-                    let entity = claim.of.clone();
-
-                    // Rehydrate the effect's source + polarity
-                    // straight off the same selection environment.
-                    let Some(effect) = load_effect_facts(&entity, env).await? else {
-                        continue;
-                    };
-
-                    let mut m = input.clone();
-                    if let Some(ref t) = this_term {
-                        m.bind(t, Value::Entity(entity.clone()))?;
-                    }
-                    if let Some(ref t) = definition_term {
-                        let definition = RuleDefinition::from_effect(&effect);
-                        let json = serde_json::to_string(&definition)
-                            .map_err(|e| EvaluationError::Store(e.to_string()))?;
-                        m.bind(t, Value::String(json))?;
-                    }
-                    yield m;
-                }
-            }
-        }
-    }
-
-    fn realize(&self, source: Match) -> Result<Self::Conclusion, EvaluationError> {
-        // `ConceptConclusion`'s fields are private; delegate to
-        // dialog's `ConceptQuery::realize`, which only reads
-        // `terms` and the match's `this` binding. The `predicate`
-        // is unused by `realize` so a stub stands in.
-        let synthetic = ConceptQuery {
-            terms: self.terms.clone(),
-            predicate: ConceptDescriptor::from(
-                Vec::<(&str, dialog_query::AttributeDescriptor)>::new(),
-            ),
-        };
-        Application::realize(&synthetic, source)
-    }
-}
-
-/// Rehydrate an effect from its `dialog.effect/source` and
-/// `dialog.effect/polarity` claims read straight off a query
-/// selection environment.
-///
-/// Equivalent to [`EffectByEntity::resolve`] but takes the raw
-/// `Provider<Select>` selection env that [`Application::evaluate`]
-/// is handed, rather than a `&Branch`. Returns `None` when the
-/// entity has no `source` claim (a dangling marker).
-async fn load_effect_facts<'a, Env>(
-    entity: &Entity,
-    env: &'a Env,
-) -> Result<Option<Effect>, EvaluationError>
-where
-    Env: Provider<Select<'a>> + Provider<SelectRules> + ConditionalSync,
-{
-    let source_claims: Vec<dialog_query::Claim> = the!("dialog.effect/source")
-        .of(Term::<Entity>::from(entity.clone()))
-        .is(Term::<String>::var("__rule_query_source"))
-        .perform(env)
-        .try_vec()
-        .await?;
-    let Some(source_claim) = source_claims.into_iter().next() else {
-        return Ok(None);
-    };
-    let Value::String(source) = source_claim.is else {
-        return Err(EvaluationError::Store(
-            "dialog.effect/source claim was not a string".to_string(),
-        ));
-    };
-
-    let polarity_claims: Vec<dialog_query::Claim> = the!("dialog.effect/polarity")
-        .of(Term::<Entity>::from(entity.clone()))
-        .is(Term::<String>::var("__rule_query_polarity"))
-        .perform(env)
-        .try_vec()
-        .await?;
-    let Some(polarity_claim) = polarity_claims.into_iter().next() else {
-        return Err(EvaluationError::Store(
-            "missing dialog.effect/polarity claim".to_string(),
-        ));
-    };
-    let Value::String(polarity_str) = polarity_claim.is else {
-        return Err(EvaluationError::Store(
-            "dialog.effect/polarity claim was not a string".to_string(),
-        ));
-    };
-    let polarity = EffectPolarity::parse(&polarity_str)
-        .ok_or_else(|| EvaluationError::Store(format!("invalid polarity {polarity_str:?}")))?;
-
-    let effect = Effect::from_source(&source, polarity)
-        .map_err(|e| EvaluationError::Store(format!("effect rehydrate failed: {e}")))?;
-    Ok(Some(effect))
-}
-
-/// The well-known descriptor for the "rule of rules" head.
-///
-/// Its `with` map names the marker (`dialog.meta/effect`) and the
-/// synthesised `definition` field — enough for the analyzer to
-/// project the fields a `rule:` head exposes. The `definition`
-/// claim has no real EAV backing; it is produced only by
-/// [`AnonymousRuleQuery::evaluate`], which the concept query
-/// planner dispatches to whenever it sees this descriptor's
-/// `this()`.
-pub fn rule_of_rule_descriptor() -> &'static ConceptDescriptor {
-    static DESCRIPTOR: std::sync::OnceLock<ConceptDescriptor> = std::sync::OnceLock::new();
-    DESCRIPTOR.get_or_init(|| {
-        serde_json::from_value(serde_json::json!({
-            "description": "Every inductive rule installed on a branch.",
-            "with": {
-                "effect":     { "the": "dialog.meta/effect",     "as": "Entity", "cardinality": "one" },
-                "definition": { "the": "dialog.effect/source",   "as": "Text",   "cardinality": "one" }
-            }
-        }))
-        .expect("rule-of-rule descriptor is well-formed")
-    })
-}
 
 /// Build a runtime [`ArtifactsAttribute`] from a domain + local
 /// name. Used by the [`assert_effect`] / [`retract_effect`]
@@ -870,69 +645,5 @@ mod tests {
             Err(EffectValidationError::MissingTrigger) => Ok(()),
             other => panic!("expected MissingTrigger, got {other:?}"),
         }
-    }
-
-    /// Install an effect on a branch, then query `rule:` via
-    /// [`AnonymousRuleQuery`] and assert it surfaces the rule with
-    /// its definition. The `definition` field round-trips to a
-    /// [`RuleDefinition`] whose `rule` descriptor and `polarity`
-    /// match the installed effect — the rule-side parallel of
-    /// `concept.rs`'s `it_returns_concept_with_source_from_concept_query`.
-    #[dialog_common::test]
-    async fn it_enumerates_installed_rules_via_rule_query() -> anyhow::Result<()> {
-        use dialog_query::{Any, Output as _, Parameters, Term};
-        use dialog_repository::helpers::{test_operator_with_profile, test_repo};
-
-        let (operator, profile) = test_operator_with_profile().await;
-        let repo = test_repo(&operator, &profile).await;
-        let branch = repo.branch("main").open().perform(&operator).await?;
-
-        // Install one assert-polarity effect on the branch.
-        let rule = InductiveRule::new(counter_head(), increment_body()).expect("rule compiles");
-        let effect = Effect::asserting(rule);
-        let effect_entity = effect.this();
-        let expected = effect.descriptor();
-        branch
-            .transaction()
-            .assert(EffectStatement(effect.clone()))
-            .commit()
-            .perform(&operator)
-            .await?;
-
-        // Query `rule:` with `this` and `definition` as variables.
-        let mut terms = Parameters::new();
-        terms.insert("this".to_string(), Term::<Any>::var("this"));
-        terms.insert("definition".to_string(), Term::<Any>::var("definition"));
-
-        let conclusions: Vec<ConceptConclusion> = branch
-            .query()
-            .select(AnonymousRuleQuery::new(terms))
-            .perform(&operator)
-            .try_vec()
-            .await?;
-
-        // Exactly one rule on the branch — the one we installed.
-        assert_eq!(
-            conclusions.len(),
-            1,
-            "expected one rule row; saw {conclusions:?}"
-        );
-        let row = &conclusions[0];
-
-        // `this` binds the effect entity.
-        let this: Entity = Entity::try_from(row.source().lookup(&Term::<Any>::var("this"))?)
-            .expect("this binding must be an entity");
-        assert_eq!(this, effect_entity);
-
-        // `definition` carries the JSON-serialised RuleDefinition.
-        let definition_json: String =
-            String::try_from(row.source().lookup(&Term::<Any>::var("definition"))?)
-                .expect("definition binding must be a string");
-        let definition: RuleDefinition = serde_json::from_str(&definition_json)?;
-        assert_eq!(definition.polarity, EffectPolarity::Assert);
-        assert_eq!(definition.rule.assert.this(), expected.assert.this());
-        assert_eq!(definition.rule.when.len(), expected.when.len());
-        assert_eq!(definition.rule.unless.len(), expected.unless.len());
-        Ok(())
     }
 }
