+++ rust/tonk-schema/src/lib.rs
@@ -47,6 +47,8 @@ pub mod concept;
 
 pub mod resolution;
 
+pub mod rule_query;
+
 pub mod effect_query;
 
 pub mod effects;
